refactor: modularised frontend file structure
