+++ client/src/main.jsx
@@ -1,13 +1,13 @@
 import App from '@/App';
 import AuthProvider from '@/contexts/AuthProvider';
+import LoginPage from '@/modules/auth/LoginPage';
+import CreateBlogPage from '@/modules/blog/pages/CreateBlogPage';
+import EditBlogPage from '@/modules/blog/pages/EditBlogPage';
+import ViewBlogPage from '@/modules/blog/pages/ViewBlogPage';
+import HomePage from '@/modules/home/HomePage';
+import SettingsPage from '@/modules/settings/SettingsPage';
+import UserBlogsPage from '@/modules/user/UserBlogsPage';
 import AboutPage from '@/pages/AboutPage';
-import CreateBlogPage from '@/pages/blogs/CreateBlogPage';
-import EditBlogPage from '@/pages/blogs/EditBlogPage';
-import ViewBlogPage from '@/pages/blogs/ViewBlogPage';
-import HomePage from '@/pages/HomePage';
-import LoginPage from '@/pages/LoginPage';
-import UserBlogsPage from '@/pages/profile/UserBlogsPage';
-import SettingsPage from '@/pages/SettingsPage';
 import { geekblue } from '@ant-design/colors';
 import { App as AntdApp, ConfigProvider } from 'antd';
 import { StrictMode } from 'react';
